LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-pt-BR-expected.txt

PASS navigator.language is "pt-br"
PASS successfullyParsed is true

TEST COMPLETE
